backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic[email]==2.5.0

# HTTP client
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# OpenAI
openai==1.3.8

# Async task queue (for future use)
celery[redis]==5.3.4
redis==5.0.1

# Monitoring and logging
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# API documentation
python-json-logger==2.0.7